add cari data buku, cari data anggota, menu utama admin, tombol back ke menu login

+++ src/form/formKelolaAnggota.java
@@ -24,6 +24,8 @@ public class formKelolaAnggota extends javax.swing.JInternalFrame {
     public formKelolaAnggota() {
         initComponents();
         load_data();
+        
+        // Btn hapus dinonaktifkan
         btnHapus.setEnabled(false);
     }
     
@@ -57,7 +59,7 @@ public class formKelolaAnggota extends javax.swing.JInternalFrame {
                 String username = rs.getString(5);
                 String status = rs.getString(7);
                 
-                // Tambahkan data dari database ke tabel
+                // Tambahkan data dari hasil query ke tabel
                 String d[] = {id_anggota, nama_anggota, no_tlp, alamat, username, status};
                 data.addRow(d);
             }
@@ -71,6 +73,8 @@ public class formKelolaAnggota extends javax.swing.JInternalFrame {
     }
     
     private void load_data_status(String stat) {
+        
+        // Mempersiapkan koneksi
         Connection kon = Koneksi.koneksiDB();
         
         // Nama header tabel
@@ -106,9 +110,59 @@ public class formKelolaAnggota extends javax.swing.JInternalFrame {
             
         } catch(Exception e) {
             JOptionPane.showMessageDialog(null, "Error: "+e);
-            formUtamaAdmin utamaAdmin = new formUtamaAdmin();
-            utamaAdmin.setVisible(true);
-            this.dispose();
+            load_data();
+        }
+    }
+    
+    private void load_data_cari() {
+        
+        // Koneksi ke database
+        Connection kon = Koneksi.koneksiDB();
+        
+        // Nama header tabel
+        Object tableHeader[] = {"ID ANGGOTA", "NAMA ANGGOTA", "NO TELEPON", "ALAMAT", "USERNAME", "STATUS"};
+        DefaultTableModel data = new DefaultTableModel(null, tableHeader);
+        
+        // Tambahkan data ke tabel Buku
+        tblAnggota.setModel(data);
+            
+        try {
+            // Mempersiapkan statement
+            Statement stmt = kon.createStatement();
+            
+            // Query sql
+            String sql = "SELECT * FROM anggota WHERE nama_anggota LIKE '%"+txtCari.getText()+"%'";
+            
+            // Eksekusi query ke database
+            ResultSet rs = stmt.executeQuery(sql);
+            
+            // Selama hasil ada ( true )
+            while(rs.next()) {
+                String id_anggota = rs.getString(1);
+                String nama_anggota = rs.getString(2);
+                String no_tlp = rs.getString(3);
+                String alamat = rs.getString(4);
+                String username = rs.getString(5);
+                String status = rs.getString(7);
+                
+                // Tambahkan data dari database ke tabel
+                String d[] = {id_anggota, nama_anggota, no_tlp, alamat, username, status};
+                data.addRow(d);
+                 
+            } 
+            
+            // Resultset 2 ketika data rs2.next = false ( data tidak ada )
+            ResultSet rs2 = stmt.executeQuery(sql);
+            
+            // Jika data tidak ada
+            if(!rs2.next()) {
+                JOptionPane.showMessageDialog(null, "Data tidak ditemukan");
+                load_data();
+            }
+            
+        } catch(Exception e) {
+            JOptionPane.showMessageDialog(null, "Error: "+e);
+            load_data();
         }
     }
     
@@ -156,6 +210,10 @@ public class formKelolaAnggota extends javax.swing.JInternalFrame {
         btnAktif = new javax.swing.JButton();
         btnNonaktif = new javax.swing.JButton();
         btnSemua = new javax.swing.JButton();
+        txtCari = new javax.swing.JTextField();
+        btnCari = new javax.swing.JButton();
+        jButton2 = new javax.swing.JButton();
+        btnKembali = new javax.swing.JButton();
 
         setPreferredSize(new java.awt.Dimension(1366, 632));
 
@@ -211,6 +269,28 @@ public class formKelolaAnggota extends javax.swing.JInternalFrame {
             }
         });
 
+        txtCari.setText("Cari nama anggota disini");
+        txtCari.addMouseListener(new java.awt.event.MouseAdapter() {
+            public void mouseClicked(java.awt.event.MouseEvent evt) {
+                txtCariMouseClicked(evt);
+            }
+        });
+
+        btnCari.setIcon(new javax.swing.ImageIcon(getClass().getResource("/image/search-icon.png"))); // NOI18N
+        btnCari.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                btnCariActionPerformed(evt);
+            }
+        });
+
+        btnKembali.setFont(new java.awt.Font("Dialog", 1, 18)); // NOI18N
+        btnKembali.setText("KEMBALI");
+        btnKembali.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                btnKembaliActionPerformed(evt);
+            }
+        });
+
         javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
         getContentPane().setLayout(layout);
         layout.setHorizontalGroup(
@@ -225,34 +305,49 @@ public class formKelolaAnggota extends javax.swing.JInternalFrame {
                         .addGap(0, 206, Short.MAX_VALUE)
                         .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
                             .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, layout.createSequentialGroup()
-                                .addComponent(btnHapus, javax.swing.GroupLayout.PREFERRED_SIZE, 157, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                .addGap(590, 590, 590))
+                                .addComponent(btnKembali, javax.swing.GroupLayout.PREFERRED_SIZE, 160, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addGap(31, 31, 31)
+                                .addComponent(btnHapus, javax.swing.GroupLayout.PREFERRED_SIZE, 160, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addGap(487, 487, 487))
                             .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, layout.createSequentialGroup()
-                                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
+                                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
                                     .addGroup(layout.createSequentialGroup()
                                         .addComponent(btnSemua)
-                                        .addGap(18, 18, 18)
+                                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
                                         .addComponent(btnAktif)
+                                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                                        .addComponent(btnNonaktif)
+                                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                                        .addComponent(jButton2, javax.swing.GroupLayout.PREFERRED_SIZE, 40, javax.swing.GroupLayout.PREFERRED_SIZE)
                                         .addGap(18, 18, 18)
-                                        .addComponent(btnNonaktif))
+                                        .addComponent(txtCari, javax.swing.GroupLayout.PREFERRED_SIZE, 202, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                                        .addComponent(btnCari, javax.swing.GroupLayout.PREFERRED_SIZE, 36, javax.swing.GroupLayout.PREFERRED_SIZE))
                                     .addComponent(jScrollPane1, javax.swing.GroupLayout.PREFERRED_SIZE, 970, javax.swing.GroupLayout.PREFERRED_SIZE))
                                 .addGap(182, 182, 182))))))
         );
         layout.setVerticalGroup(
             layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(layout.createSequentialGroup()
-                .addGap(30, 30, 30)
-                .addComponent(jLabel2, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addGap(19, 19, 19)
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(btnAktif)
-                    .addComponent(btnNonaktif)
-                    .addComponent(btnSemua))
+                .addComponent(jLabel2, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(45, 45, 45)
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
+                    .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                        .addComponent(btnSemua, javax.swing.GroupLayout.PREFERRED_SIZE, 36, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addComponent(btnAktif, javax.swing.GroupLayout.PREFERRED_SIZE, 36, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addComponent(btnNonaktif, javax.swing.GroupLayout.PREFERRED_SIZE, 36, javax.swing.GroupLayout.PREFERRED_SIZE))
+                    .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                        .addComponent(txtCari, javax.swing.GroupLayout.PREFERRED_SIZE, 36, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addComponent(jButton2, javax.swing.GroupLayout.PREFERRED_SIZE, 36, javax.swing.GroupLayout.PREFERRED_SIZE))
+                    .addComponent(btnCari, javax.swing.GroupLayout.PREFERRED_SIZE, 36, javax.swing.GroupLayout.PREFERRED_SIZE))
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                .addComponent(jScrollPane1, javax.swing.GroupLayout.PREFERRED_SIZE, 360, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(jScrollPane1, javax.swing.GroupLayout.PREFERRED_SIZE, 375, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(18, 18, 18)
-                .addComponent(btnHapus, javax.swing.GroupLayout.PREFERRED_SIZE, 68, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap(25, Short.MAX_VALUE))
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(btnKembali, javax.swing.GroupLayout.PREFERRED_SIZE, 68, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(btnHapus, javax.swing.GroupLayout.PREFERRED_SIZE, 68, javax.swing.GroupLayout.PREFERRED_SIZE))
+                .addContainerGap(28, Short.MAX_VALUE))
         );
 
         pack();
@@ -266,8 +361,15 @@ public class formKelolaAnggota extends javax.swing.JInternalFrame {
     }//GEN-LAST:event_tblAnggotaMouseClicked
 
     private void btnHapusActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnHapusActionPerformed
-        hapus_data();    
-        load_data();
+        int hapus = JOptionPane.showOptionDialog(this, 
+                "Apakah anda yakin untuk menghapus data ini ?", null, JOptionPane.YES_NO_OPTION,
+                JOptionPane.QUESTION_MESSAGE, null, null, null);
+        
+        if(hapus == JOptionPane.YES_OPTION) {
+            hapus_data();
+            load_data();
+            btnHapus.setEnabled(false);
+        }
     }//GEN-LAST:event_btnHapusActionPerformed
 
     private void btnAktifActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnAktifActionPerformed
@@ -284,14 +386,30 @@ public class formKelolaAnggota extends javax.swing.JInternalFrame {
         load_data_status("TIDAK AKTIF");
     }//GEN-LAST:event_btnNonaktifActionPerformed
 
+    private void btnCariActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnCariActionPerformed
+        load_data_cari();
+    }//GEN-LAST:event_btnCariActionPerformed
+
+    private void btnKembaliActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnKembaliActionPerformed
+        this.dispose();
+    }//GEN-LAST:event_btnKembaliActionPerformed
+
+    private void txtCariMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_txtCariMouseClicked
+        txtCari.setText("");
+    }//GEN-LAST:event_txtCariMouseClicked
+
 
     // Variables declaration - do not modify//GEN-BEGIN:variables
     private javax.swing.JButton btnAktif;
+    private javax.swing.JButton btnCari;
     private javax.swing.JButton btnHapus;
+    private javax.swing.JButton btnKembali;
     private javax.swing.JButton btnNonaktif;
     private javax.swing.JButton btnSemua;
+    private javax.swing.JButton jButton2;
     private javax.swing.JLabel jLabel2;
     private javax.swing.JScrollPane jScrollPane1;
     private javax.swing.JTable tblAnggota;
+    private javax.swing.JTextField txtCari;
     // End of variables declaration//GEN-END:variables
 }
